Dual normalization (#75)
* Better dual docstring

* Converting to A-SOH in degradation wrapper

* Better way of converting hidden to A-SOH in degradation wrapper

* Streamlining way to get states

Create a separate function to retrieve states from filters and convert their
values to model coordinate system (but they are returned as if they were filter-
related objects). This functionality was needed in a number of places, so it made
sense to make it separate and easily called.

* Allow MvRD to be converted through inverse

Adding a keyword to the convert method allows the MvRD to be inverted, as well
as converted, without the need to create an additional ConversionOperator.

* Fixing bug in joint step

Inputs and measurements needed to be inverted into filter lingo before stepping
the underlying filter. That has now been rectified.

* Easier way to perform joint step

* Test inverted conversions

* Fixes to normalization in dual estimation

A number of fixes were needed, including in the initialization of the dual UKF
with appropriate hidden states, noise covariances, etc. Additionally, a few
fixes were also needed to make sure normalization was properly applied from one
filter to the other in the appropriate copies.

* Adjusting DegradationWrapper

Update hidden state now assembles a true A-SOH object and reconverts it, which
should make it easier when we have a real degradation model.

* Adding test to account for normalization of dual

One more test that runs an A-SOH normalized dual UKF. For some reason, it
requires one more cycle than the un-normalized one to converge...

* Making sure Gaussian mean is provided flattened

* New simple test on dual

Makes sure a simple transformation works

* Avoiding unnecessary copy

* Test proper initialization with A-SOH normalization