Nonlinear transform (#78)
* Adding non linear conversion operators

Added a base class to deal with non-linear conversion operators with uncertainty
(covariance) propagation based on a first-order Taylor series expansion of the
underlying transformation. Also create a child of that class for converting to
absolute values.

* Testing AbsoluteValue conversion

This conversion is tested here. It should be particularly useful for the ECM
A-SOH case, where the UKF tends to want to sample illegal regions of space.

* Adding pivot to all covariance transforms

* Adding pivots to MultivariateGaussian

For cases where the conversion operators rely on Taylor-expansions

* Removing Optional flags

The converters cannot be None, therefore, they should not take the Optional flag.
The fact if has a default value does not mean it is an optional argument.

* Absolute value can now be applied to specific indices

* Minor fixes to joint UKF init

* Minor corrections to dual UKF init

* Rerunning dual UKF Thenevin with AbsValConversion

Using the new AbsoluteValueConversionOperator, we are able to step the dual UKF
for the Thenevin example all the way through, without sampling illegal A-SOH
regions. However, the predictions of the dual estimator remain very poor.

* Addressing comments

Better readability and efficiency

* Docstring fix