Rerun with SOC independent HealthVariables
From tests with the dual estimation, it appears that, when the health variables
being updated are SOC independent (with the notable exception of the hysteresis
parameter H0!), the dual estimation works fine, but it takes much longer to
converge than its joint counterpart. However, when any of the health variables
changes value based on SOC (such as R0 having several base values), the dual
estimator fails to capture the ground truth: instead, it keeps the SOC fixed at
some level, and changes the values of the parameters at that level. At times, it
gives completely unphysical results, such as Qt that is orders of magnitude
apart from the real Qt. In my opinion, maybe a denoiser could help.