Fixing bug on interpolator
Found a bug while testing the degradation model with the joint filter. In short,
calls to the ECM.update_transient_state where altering the value of the base
values of the H0 attribute of the A-SOH. After some investigation, I found out
it was happening in the special case where interpolation was not necessary: we
needed to copy the base values, otherwise, we would be changing the base_values
in the H0. That seems to have fixed the issue.