Fixing bug in interpolation
The SOCInterpolation would add a dimension when we didn't want it to. In the case
where the SOC was passed as a 2D array of shape (1,1), and the health variable
batch size was N, the returned array would be of shape (N,1,1), which is not the
desired effect. Instead, now, we only return a 2D array of shape (N,1) in these
cases, and, when the SOC is passed as a 0D or 1D array, we return it in those
shapes as well.