Degradation model (#85)
* Creating base DegradationModel

Base class to establish the minimum requirements of a degradation model. Children
of this class maybe need additional inputs, and may need to store previous states
as well.

* Creating dummy degradation model

Degradation model that does not degrade the A-SOH.

* Allowing transient history to be passed to degradation model

* Adding test for dummy degradation

* minor modifications to base degradation

* Allowing outputs to be passed to degradation as well

* Fixing model tests

We now require CellModels to be passed to the simulator with the 'cell_model'
keyword argument (or as a positional argument too), so some tests had to be
fixed.

* Adding degradation model to simulator

* Fixing estimator tests

The Simulator no longer takes a 'model' keyword argument, instead it takes a
'cell_model' one.

* Fixing notebooks

Since simulator now requires 'cell_model' instead of 'model', notebooks had to
be fixed.

* Adding degradation to ModelWrappers

Both dual and joint model wrappers now have degradation.

* Fixing bug on interpolator

Found a bug while testing the degradation model with the joint filter. In short,
calls to the ECM.update_transient_state where altering the value of the base
values of the H0 attribute of the A-SOH. After some investigation, I found out
it was happening in the special case where interpolation was not necessary: we
needed to copy the base values, otherwise, we would be changing the base_values
in the H0. That seems to have fixed the issue.

* rerunning joint demo after degradation changes

* Making sure degradation works with joint estimation

* fixing typo

* Fixing bug in interpolation

The SOCInterpolation would add a dimension when we didn't want it to. In the case
where the SOC was passed as a 2D array of shape (1,1), and the health variable
batch size was N, the returned array would be of shape (N,1,1), which is not the
desired effect. Instead, now, we only return a 2D array of shape (N,1) in these
cases, and, when the SOC is passed as a 0D or 1D array, we return it in those
shapes as well.

* Rerunning demos to ensure all is ok

* Addressing comments

* Testing interpolation when SOC batch size is 1