Revamped batching when interpolating
We have slightly altered the way batching is done in the SOCInterpolatedHealth.
The main array it works with now is 3D, having the internal batch size as the
first dimension, the SOC batch size as the second, and the SOC dimensionality as
the third. It takes an additional flag that allows the user to get this array
returned, but it's default configuration is to try to match the internal and SOC
batches if possible, as well as to remove superfluous batches of cardinality one.