Reworking batching interpolation (#96)
* Rerunning notebooks after new changes

Making sure they all still at least run, and, luckily, still produce the same
results!

* Revamped batching when interpolating

We have slightly altered the way batching is done in the SOCInterpolatedHealth.
The main array it works with now is 3D, having the internal batch size as the
first dimension, the SOC batch size as the second, and the SOC dimensionality as
the third. It takes an additional flag that allows the user to get this array
returned, but it's default configuration is to try to match the internal and SOC
batches if possible, as well as to remove superfluous batches of cardinality one.

* Adding more explanation

Added documentation with more detailed explanation of how batching works in the
SOCInterpolatedHealth variable, including explicit mentions of the shape of
arrays used and returned, and the impacts of the boolean flag.

* Changing return from SOCInterpolatedHealth

Now, SOCInterpolatedHealth get_value method always returns a 3D array with
shape (internal_batch, soc_batch, soc_dim)

* Altered some computations

Had to change the way some things are computed to accomodate the changes to the
SOCInterpolatedHealth class, which now always returns a 3D array from its
get_value method. Also included comments detailing the shapes of most arrays
being used, so we don't get lost.

* Removed flags that are no longer used

Everything in here was a child of SOCInterpolatedHealth, which no longer uses
an optional flag for broadcasting batches together.

* Additional tests on the components of an ECM

* removing leftover print statements

* Making sure OCV returns correct dimensions

Since we changed the dimensionality of the base get_value method, we should
also make sure the OCV calls return a 3D array for consistency.

* Additional changes to dimensionality

Because of the new 3D array returns, there were a few additional bugs to be
ironed out, including:
1. the summation of the overpotential in the RC elements was keeping an unwanted
dimension
2. the computation of the RC time constants creating an array of the wrong shape
due to a squeeze followed by a at_least2d, which messed up the batch dimension
3. not sure this was a problem, but a similar situation of the squeeze was used
in the hysteresis, so I removed that just to be sure

* Rerunning notebooks to make sure everything works