Energy calculation in ECMASOH objects (#99)
* Adding energy calculation to ECMASOH

Now, as part of the ECMASOH object, we can easily get the maximum theoretical
energy available.

* Starting tests for energy calculation

We need to test the energy calculation function, which seems to still be broken
under batching.

* Qt amp_hour broken when batched

Because the amp-hour functionality was returning a .item() from the base values,
when that object is batched, the method would fail. Now, we just simply return
the base values.

* Checking batch size

Just trying something out. Not yet complete, but come back to this later.

* ECMASOH can compute its own energy

The ECMASOH objects now come with the ability to computed the energy available as
a function of temperature and the desired SOC limits. This computation also works
when the ECMASOH object is batched. It always returns a 2D array with shape
(batch_size, 1)

* Adding test for shape of energy array

* Better phrasing in the documentation.

Avoided using first-person language and kept the documentation wording more
impersonal.