Test did not account for hysteresis
In one of the tests, we calculate the expected terminal voltage by hand, but we
did not account for hysteresis, as that was not distinguishable from a floating
point error at the time we created this test. That is no longer the case, so we
had to manually set the max hysteresis parameter to 0.