Adds log-likelihood to distributions (#140)
* ignoring vscode settings

* MV Random Distributions must compute log likelihood

All children of MultivariateRandomDistribution must be able to compute the log
likelihood of data (output as one log-likelihood per datapoint provided, so not
of the full data). This has been implemented for Delta and Gaussian distributions

* testing new capability

* Small adjustments to get rid of warnings

Some things return numpy arrays which are fed to other things that require float,
so that was fixed with .item(). Similarly, got rid of a deprecation warning from
pydantic, as we were using copy rather than model_copy

* Improvements to log likelihood computations

Fixing typos, avoiding floating point issues, and using established packages when
possible.