Changing solver for I_RC
Before, we were using an explicit Euler stepping technique to solve for the
current at the resistive element. This was prone to issues when the time step
between observations during the non-rest period is high. An alternative to help
dealing with that is to use an implicit Euler method, which, luckily for us, can
be easily expressed analytically. This brings better stability to the solutions