More general RCExtractor (#172)
* Reorganizing Extractors

ECM-based extractors are now reorganized, so that it is easier to make
modifications and improvements to individual components without having to deal
with a large file. All tests are passing as well.

* Extracti_rests given I(t) info from previous step

Since we will need to compute the current flowing through the resistive element
at the end of the previous step, we need to know the full history of total
current vs. time for this step. We now return that in the form of a dictionary

* Function to compute the i_RC at the end of a step

* making sure we test for when there is a charge in the capacitor

* small modifications to make sure tests still pass

* printing to check out what is wrong

* Changing solver for I_RC

Before, we were using an explicit Euler stepping technique to solve for the
current at the resistive element. This was prone to issues when the time step
between observations during the non-rest period is high. An alternative to help
dealing with that is to use an implicit Euler method, which, luckily for us, can
be easily expressed analytically. This brings better stability to the solutions

* Adding relative tolerance to tests

Due to the change to an implicit solver, we had to add a relative tolerance to
one of the tests, but we are still making sure the solution is correct (tolerance
is of 0.1%)

* RCExtractor now uses better guess for I_RC

This makes sure that estimating the resistance in the RC component yields better
results, even when the previous non-rest step is not long enough to saturate the
RC component.

* Computing R and C with better current estimate

Now that we have a more reliable estimate of the current that flows through the
resistive element of the RC component at the beginning of the rest period, we
can more accurately compute the values of R and C from the fitted A and tau
parameters.

* Testing improved RCExtractor

Including some more realistic HPPC data in which the RC component is not fully
saturated to make sure the current computation works as intended.

* removing print statements from tests

* fixing typo with variable name

* better documentation