Quick capacity extraction fix (#174)
* Fixing issue with capacity extraction

Previously, the capacity extraction relied solely on the `'capacity_charge'` flag
from the battery data toolkit. However, that assumes that the cell started from a
discharge state, and that it started and ended at the same SOC level. New
additions to the toolkit also include the computation of the maximum amount of
cycled charge, which is what we are after here. By leveraging this, we can obtain
better estimates of the capacity, which are independent of the starting and final
SOC during the cycle, as well as the first step (charge or discharge)

* including capacity extraction in the realistic RPT test

* checking for presence of the correct column