Assembling refiners
What was previously just considered offline estimation is now what we are
calling refiners. We are changing offline estimation to have different types of
components. For now, they are extractors, which look at raw data and extract
relevant aSOH values if they can, and refiners, which use the extracted values
as initial guesses for optimization routines.