Offline rework (#181)
* starting reorganization

* fixing documentation notebooks

* more fixes to documentation notebook

* rerunning locally to make sure it works

* Assembling refiners

What was previously just considered offline estimation is now what we are
calling refiners. We are changing offline estimation to have different types of
components. For now, they are extractors, which look at raw data and extract
relevant aSOH values if they can, and refiners, which use the extracted values
as initial guesses for optimization routines.

* fixing demo notebook

* fixing thevenin test to reflect changes on refiners

* utility function to always deal with a BatteryDataset

* defining separate data checker classes, which can be shared among different extractors

* First pass on a capacity check data checker

Here, we implement a capacity check data checker, which checks if the data
provided has segments that reach the voltage limits, and which are slow enough
to be considered capacity checks.

* reorganization of data checkers

* adding unit test for capacity check data checker

* Extractors can now receive raw data or Dataset object

* fixing line that is too long

* stylistic fixes

* Creating separate single cycle and SOC range checkers

Now, we have a checker to make sure the data is from a single cycle, as well as
one to make sure the approximate range of SOC values sampled is adequate. That
still needs testing.

* Reorganizing simple DataCheckers and testing SOC checker

* Adding HPPC check and testing it

Making sure we can identify pulses, and double check if they are somewhat
bidirectional. We do so by just counting the number of charge and discharge
pulses, but really, we should double check they follow one another closely at
similar SOC levels.

* fixing some linting in the tests

* more linting fixes

* fixing conftest blank line with spaces

* more style fixes

* changing nomenclature

* allowing pulses to be returned if desired

* adding a checker for rest periods

* linting fixes

* Reflecting changes from the toolkit into extractors

Some of the column names have changed in the BatteryDataToolkit, and these
changes are reflected here.

* Including Coulombic Efficiency considerations

Since now we can take Coulombic Efficiency into account in the Battery Data
Toolkit, we should do the same here whenever possible.

* including full HPPC checker

* style fix

* more style fix

* linting issues

* Data checkers now return postprocessed data

Since data checkers may need to perform postprocessing on the data, such as
tagging and coulombic counting calculations, it makes more sense if these are
stored with the data object and returned, so they don't have to be re-done.

* linting fixes

* Reworking capacity and CE Extractor

Now, capacity and CE extractor follow the new format.

* fixing flake8 issues

* better error message for C-rates larger than 1

* reflecting changes on BatteryDataToolkit retrieval of strings

* Modularizing capacity

This should make it easier for a user to bypass checks if desired.

* allowing for further flexibility in usage of capacity extractor

* changes to test names for brevity

* adding the ability to extract values without checks in base

* reflecting changes in base here

* typo fix

* Creating series resistance extractor

Also using additional helper functions where needed

* Adding helper properties

This makes it easier to access things like capacity and coulombic efficiency

* reflecting changes in battery data toolkit nomenclature

* tests for R0 extractor

* adding condition on duration of steps that precede rest for rest data checker

* Utility functions for fitting exponential decays

Making it easier to fit exponential decay functions for RC parameters, as well
as easier to test them.

* minor changes to series resistance properties

* ensuring compatibility with BatteryDataToolkit

* Removing offset from measurements

Since we are only interested in fitting an exponential decay that reaches its
stable configuration, we can safely remove the offset from the measurements when
fitting.

* minor style fix

* allowing offset to be fit as well, but ignoring it

* RC Extractor seemingly complete

RC extractor has been completed, and seems to be working as intended.

* linting fixes

* OCV extractor done

We have now a new version of the OCV extractor implemented

* Fixing issues with the documentation notebook

The changes to the extractors needed to be reflected on this notebook. Note,
however, that this notebook is very outdated and needs to be fully re-done.

* additional info returned from OCV extraction

* creating assemblers

* creating capacity assembler

* Creating utility functions for ECM Assemblers

This should make it easier to create a unified form of assembler that simply
uses different forms of SOC fits.

* reorganizing SOC regressor

* creating general SOC-dependent assembler

* first test for assemblers

* testing capacity

* better capacity tests

* testing OCV assembler

* resistance assembler complete

* Capacitance assembler done

* style fixes

* nomenclature change to refiners

* Loss definition altered

Now, the observations are only used at call time, so that we can easily apply the
same loss algorithm to different datasets

* changes on losses now reflected on refiners

* style fix

* fixing test given new changes in refiners

* adding dictionary to help initialize FullHPPCDataChecker with more ease

* params to pre-initialize R0Extractor

* Starting framework for Offline Estimation

We just started the framework for offline estimation of ECM parameters from RPT
data. Still need to refine the final result (that is, extraction and assembly
are done, but optimization step still missing)

* adding refiner to the offline estimator

* making sure returned aSOH is all fixed for consistency

* More customization to OfflineEstimator

Now, the SOC-pinpoints for hysteresis can also be provided. We might want to
consider instead just having a hysteresis extractor. Additionally, we included
the option to specify which parameters one would like to refine, as well as
not refining any parameters.

* creating simple hysteresis extractor

* testing H0 extraction

* computing weights for assemblers the right way

* better extraction for weights

* Better way to extract hysteresis

Now, to extract the instantaneous hysteresis, we simulate the cell without any
hysteresis and compare the difference.

* hysteresis extractor returns signed values

* Better hysteresis assembler

With the improved hysteresis extractor, we also needed to change a few details
on the hysteresis assembler. It is far from perfect, but appears to be somewhat
serviceable.

* adding test to offline estimator framework

* style fixes

* fixing small issue with dual estimator results

* introducing numpy polynomial fit to SOC regressor

* checking for negative values in hysteresis

* starting workflow demo

* adding offline portion of demo workflow

* style fix

* Workflow demonstration notebook complete

The notebook includes everything: how to generate data, how to perform offline
estimation from RPT, and how to set up and run online estimation.